Make SettingManger as static class
:Release Notes:
In the previous implementation, the distinction between Singleton and
Static class was not clear.
As the first work to clearly distinguish Singleton and Static class,
make SettingManager as static class.

SettingManager class is stateless. Also it is referenced globaly.
It doesn't have any hierarchycal dependencies with other class.
It is structually good to make SettingManger as static class.

:Detailed Notes:
- Make SettingManager as static class
- Refactoring SettingManager class
- Refactoring methods which use SettingManager

:Testing Performed:
QWA-347 : OK
QWA-346 : OK
QWA-351 : OK
QWA-355 : OK
QWA-2391 : OK

:QA Notes:
N/A

:Issues Addressed:
[PLAT-120937] Refactoring interface and base class

Change-Id: I927005ad8b6c62b211187952dd318c2e4a8b7f2e